+++ Source/Sample.cs
@@ -13,7 +13,7 @@ namespace RimThreaded
                 __instance.source.pitch = __instance.SanitizedPitch;
                 __instance.ApplyMappedParameters();
                 __instance.source.volume = __instance.SanitizedVolume;
-                __instance.source.mute = (double)__instance.source.volume < 1.0 / 1000.0;
+                __instance.source.mute = __instance.source.volume < 1.0 / 1000.0;
                 if (!__instance.subDef.tempoAffectedByGameSpeed || __instance.Info.testPlay)
                     return false;
                 if (Current.ProgramState == ProgramState.Playing && Find.TickManager.Paused)
